Move filter management to separate module
